+++ packages/react-native-sdk/expo-config-plugin/__tests__/withAndroidPermissions.test.ts
@@ -26,7 +26,7 @@ describe('withStreamVideoReactNativeSDKAndroidPermissions', () => {
       expect.arrayContaining([
         'android.permission.POST_NOTIFICATIONS',
         'android.permission.FOREGROUND_SERVICE',
-        'android.permission.FOREGROUND_SERVICE_MICROPHONE',
+        'android.permission.FOREGROUND_SERVICE_DATA_SYNC',
         'android.permission.FOREGROUND_SERVICE_MEDIA_PROJECTION',
         'android.permission.BLUETOOTH',
         'android.permission.BLUETOOTH_CONNECT',