+++ sample-apps/react/messenger-clone/src/components/ChannelHeader/ChannelHeader.tsx
@@ -47,8 +47,6 @@ const UnMemoizedChannelHeader = (props: ChannelHeaderProps) => {
       <Avatar
         image={displayImage}
         name={displayTitle}
-        shape="rounded"
-        size={channel?.type === 'commerce' ? 60 : 40}
       />
       <div className="str-chat__header-livestream-left str-chat__channel-header-end">
         <p className="str-chat__header-livestream-left--title str-chat__channel-header-title">